BugFix: fixed check PageShift param in functions
All functions did not correctly check the PageShift parameter.

Page == MEM_PAGE_SIZE == 256 bytes.
Offsets (PageShift) starts from zero.

So a valid offset (PageShift)value is:
for reading/writing:
bytes: [0..255]
words: [0..254]
long words: [0..252]

When writing arbitrary data with a length == Len,
the offset must be in the range [0..256-len]

This patch fixes the code so that the offset
check is done according to the above table.